Don't save resources with no content. This fixes a bug where urlfetches were failing, causing null contents to be cached such that they were not refetched.
A more elegant solution is probably necessary here in which "permanent" errors
like 404 are cached and temporary errors like an unreachable host are cached
with a timeout.

However, this change is still necessary as resources with null contents are
almost certainly always an internal error and so should not be cached.